Removes "redundant" data points from dataset
More than one spectrum might be captured for a particular galaxy alone. In order to clean the dataset of this redundancy, we consolidate the data points within 5 arcsec of one another and then choose one data point based on a pre-defined priority order (in terms of database).

Other nuances that are used to remove certain data points have not been delved into.